Fixed logic error in command answer handler setup.
Even though the answer handler for genmove made a command to the engine
and set up an answer handler, as soon as we returned from the answer
handler for the first command, we set self.command_answer_handler to None.

Now, the answer handler itself marks that it has answered the message by
setting self.command_answer_handler to None.  It can then make another
command and things won't interfere with eachother.