WIP Fix it so it runs
I don't know what it is but although it ran before, now the relative
imports didn't work so I had to change all of those and there were a
couple other things that I had to change.

Now I have a functional board getting drawn when running

    python3 <this_dir>/python/igo/controller.py

but this thing is due for a complete rewrite in a statically typed
language so I can work on it without risking breaking anything.